requirement.txt
# LangGraph + LangChain
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0

# OpenAI
openai>=1.30.0

# Web server
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0

# Data validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0